split nat and uni assay into separate files
