+++ more/peg/peg.js
@@ -1,5 +1,5 @@
 import harden from '@agoric/harden';
-import { makeNatAssay } from '../../core/config/assays';
+import { makeNatAssay } from '../../core/config/natAssay';
 import { makeMint } from '../../core/issuers';
 
 // This implementation may be out of date
